requirements.txt
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
black==26.3.1
build==1.4.3
CacheControl==0.14.4
certifi==2026.2.25
charset-normalizer==3.4.7
cleo==2.1.0
click==8.3.2
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
Django==6.0.4
django-filter==25.2
django-phonenumber-field==8.4.0
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
drf-spectacular==0.29.0
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.28.0
findpython==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
librt==0.9.0
more-itertools==11.0.2
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.4
pbs-installer==2026.4.7
phonenumbers==9.0.28
pkginfo==1.12.1.2
platformdirs==4.9.6
poetry-core==2.3.2
pyproject_hooks==1.2.0
python-discovery==1.2.2
python-dotenv==1.2.2
pytokens==0.4.1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.5
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
shellingham==1.5.4
sqlparse==0.5.5
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==21.2.4
zstandard==0.25.0